kill monsters
This implements just enough of a comabt system to be able to kill
monsters, they currently don't attempt to attack you though so it's
very one-sided at the moment. I will (eventually?) need to fix that
along with some kind of health re-gen system I suppose.

This is however a functional if _very_ bare bones game.